requirements.txt
# Core dependencies
streamlit==1.29.0
langgraph==0.2.50
langchain==0.3.10
langchain-openai==0.2.10
langchain-core==0.3.20

# Additional utilities
python-dotenv==1.0.0
typing-extensions==4.8.0

# For better error handling and logging
loguru==0.7.2

# For session state management
uuid==1.30
